PingChat-backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
dnspython==2.7.0
eventlet==0.40.1
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
